Added example for multi image
