+++ OfficeIMO.Examples/OfficeIMO.Examples.csproj
@@ -11,6 +11,15 @@
   </ItemGroup>
 
   <ItemGroup>
+    <None Update="Images\example.gif">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="Images\saturn.tif">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
+    <None Update="Images\snail.bmp">
+      <CopyToOutputDirectory>Always</CopyToOutputDirectory>
+    </None>
     <None Update="Templates\BasicDocumentWithBookmarks.docx">
       <CopyToOutputDirectory>Always</CopyToOutputDirectory>
     </None>
